requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-google-genai
chromadb
pypdf
sentence-transformers
watchdog
